Fix bug: Create results directory
Update code to create results directory when it does not exist

+++ helper.py
@@ -2,6 +2,7 @@ import pickle
 
 import joblib
 import orjson
+import os
 import pandas as pd
 from imblearn.combine import SMOTETomek
 from imblearn.ensemble import BalancedRandomForestClassifier
@@ -185,7 +186,11 @@ def encoder(data, method="train"):
         )
 
         # Creates a joblib file for future encoding of test set
-        joblib.dump(oe, "./results/nucleotide_encoder.joblib")
+        if not os.path.exists("./results"):
+            os.makedirs("./results")
+            joblib.dump(oe, "./results/nucleotide_encoder.joblib")
+        else:
+            joblib.dump(oe, "./results/nucleotide_encoder.joblib")
 
         return train
